Check if sddm user exists
Summary:
Sddm-kcm should depend on sddm.
Sddm sets up an sddm user which is used to run the greeter.
Therefore it should not be possible to run sddm-kcm without an sddm user existing.

But if it should happen that's the case, this patch adds a check for the sddm user.
If the user is not present, everything will be aborted and an error message will pop-up.

openSUSE BUG:
  - https://bugzilla.opensuse.org/show_bug.cgi?id=1145182
  - https://bugzilla.opensuse.org/show_bug.cgi?id=1146817
  - https://bugzilla.opensuse.org/show_bug.cgi?id=1146831

Test Plan:
changed line `QString sddmHomeDirPath = KUser("sddm").homeDir();` to look up some bogus user
syncing and resetting is aborted, error message pops up

{F7273960}

Reviewers: #plasma, davidedmundson, ngraham, #vdg, fvogt

Reviewed By: ngraham, #vdg

Subscribers: davidedmundson, plasma-devel

Tags: #plasma

Differential Revision: https://phabricator.kde.org/D23378

+++ sddmauthhelper.cpp
@@ -67,8 +67,17 @@ void SddmAuthHelper::copyFile(const QString &source, const QString &destination)
 
 ActionReply SddmAuthHelper::sync(const QVariantMap &args)
 {
+    // initial check for sddm user; abort if user not present
+    // we have to check with QString and isEmpty() instead of QDir and exists() because
+    // QDir returns "." and true for exists() in the case of a non-existent user;
+    QString sddmHomeDirPath = KUser("sddm").homeDir();
+    if (sddmHomeDirPath.isEmpty()) {
+        qDebug() << "Cannot proceed, user 'sddm' does not exist. Please check your SDDM install.";
+        return ActionReply::HelperErrorReply();
+    }
+
     // create SDDM config directory if it does not exist
-    QDir sddmConfigLocation(args[QStringLiteral("sddmUserConfig")].toString());
+    QDir sddmConfigLocation(sddmHomeDirPath + QStringLiteral("/.config"));
     if (!sddmConfigLocation.exists()) {
         QDir().mkpath(sddmConfigLocation.path());
     }
@@ -151,7 +160,16 @@ ActionReply SddmAuthHelper::sync(const QVariantMap &args)
 
 ActionReply SddmAuthHelper::reset(const QVariantMap &args)
 {
-    QDir sddmConfigLocation(args[QStringLiteral("sddmUserConfig")].toString());
+    // initial check for sddm user; abort if user not present
+    // we have to check with QString and isEmpty() instead of QDir and exists() because
+    // QDir returns "." and true for exists() in the case of a non-existent user;
+    QString sddmHomeDirPath = KUser("sddm").homeDir();
+    if (sddmHomeDirPath.isEmpty()) {
+        qDebug() << "Cannot proceed, user 'sddm' does not exist. Please check your SDDM install.";
+        return ActionReply::HelperErrorReply();
+    }
+
+    QDir sddmConfigLocation(sddmHomeDirPath + QStringLiteral("/.config"));
     QDir fontconfigDir(args[QStringLiteral("sddmUserConfig")].toString() + QStringLiteral("/fontconfig"));
 
     fontconfigDir.removeRecursively();
